+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # node-red-contrib-viveresmartgroup-knxultil
 
 <p>
+<b>Version 0.0.4</b><br/>
+- Passthrough original input message to the output message.</br>
+</p>
+<p>
 <b>Version 0.0.3</b><br/>
 - Fixed output not working.</br>
 </p>